only created user can del post

+++ client/components/Post.jsx
@@ -9,11 +9,13 @@ const Post = (props) => {
   const [savedPost, setSavedPost] = useState(false)
 
   const { getPosts } = props
-  const { id, body, created_at: createdAt, first_name: firstName, last_name: lastName } = props.post
+  const { id, body, created_at: createdAt, first_name: firstName, last_name: lastName, author_id: authorId } = props.post
 
   const delPost = () => {
-    deletePost(id)
-    getPosts()
+    if (authorId === props.user.id) {
+      deletePost(id)
+      getPosts()
+    }
   }
 
   useEffect(() => {
